requirements.txt
Pillow==9.5.0
Django==3.2.23
djangorestframework==3.14.0
asgiref==3.7.2
sqlparse==0.4.4
typing-extensions==4.7.1